requirements-dev.txt
yfinance
pandas
holidays
apache-airflow==2.10.2
